cleanup and bug fixes
