+++ invokeai/frontend/web/public/locales/en.json
@@ -20,6 +20,7 @@
         "archiveBoard": "Archive Board",
         "archived": "Archived",
         "autoAddBoard": "Auto-Add Board",
+        "boards": "Boards",
         "bottomMessage": "Deleting this board and its images will reset any features currently using them.",
         "cancel": "Cancel",
         "changeBoard": "Change Board",
@@ -35,8 +36,10 @@
         "movingImagesToBoard_other": "Moving {{count}} images to board:",
         "myBoard": "My Board",
         "noMatching": "No matching Boards",
+        "private": "Private",
         "searchBoard": "Search Boards...",
         "selectBoard": "Select a Board",
+        "shared": "Shared",
         "topMessage": "This board contains images used in the following features:",
         "unarchiveBoard": "Unarchive Board",
         "uncategorized": "Uncategorized",
