+++ invokeai/frontend/web/src/features/gallery/components/Boards/BoardsList/AddBoardButton.tsx
@@ -4,13 +4,17 @@ import { useTranslation } from 'react-i18next';
 import { PiPlusBold } from 'react-icons/pi';
 import { useCreateBoardMutation } from 'services/api/endpoints/boards';
 
-const AddBoardButton = () => {
+type Props = {
+  privateBoard: boolean;
+};
+
+const AddBoardButton = ({ privateBoard }: Props) => {
   const { t } = useTranslation();
   const [createBoard, { isLoading }] = useCreateBoardMutation();
   const DEFAULT_BOARD_NAME = t('boards.myBoard');
   const handleCreateBoard = useCallback(() => {
-    createBoard(DEFAULT_BOARD_NAME);
-  }, [createBoard, DEFAULT_BOARD_NAME]);
+    createBoard({ board_name: DEFAULT_BOARD_NAME, private_board: privateBoard });
+  }, [createBoard, DEFAULT_BOARD_NAME, privateBoard]);
 
   return (
     <IconButton
@@ -19,8 +23,9 @@ const AddBoardButton = () => {
       tooltip={t('boards.addBoard')}
       aria-label={t('boards.addBoard')}
       onClick={handleCreateBoard}
-      size="sm"
+      size="md"
       data-testid="add-board-button"
+      variant="ghost"
     />
   );
 };
