+++ invokeai/frontend/web/src/features/gallery/components/ImageGalleryContent.tsx
@@ -9,7 +9,6 @@ import { PiImagesBold } from 'react-icons/pi';
 import { RiServerLine } from 'react-icons/ri';
 
 import BoardsList from './Boards/BoardsList/BoardsList';
-import BoardsListWithPrivate from './Boards/BoardsListWithPrivate/BoardsListWithPrivate';
 import GalleryBoardName from './GalleryBoardName';
 import GallerySettingsPopover from './GallerySettingsPopover/GallerySettingsPopover';
 import GalleryImageGrid from './ImageGrid/GalleryImageGrid';
@@ -19,7 +18,6 @@ import { GallerySearch } from './ImageGrid/GallerySearch';
 const ImageGalleryContent = () => {
   const { t } = useTranslation();
   const galleryView = useAppSelector((s) => s.gallery.galleryView);
-  const allowPrivateBoards = useAppSelector((s) => s.config.allowPrivateBoards);
   const dispatch = useAppDispatch();
   const galleryHeader = useStore($galleryHeader);
   const { isOpen: isBoardListOpen, onToggle: onToggleBoardList } = useDisclosure({ defaultIsOpen: true });
@@ -44,21 +42,15 @@ const ImageGalleryContent = () => {
       gap={2}
     >
       {galleryHeader}
-      {true ? (
+      <Box>
+        <Flex alignItems="center" justifyContent="space-between" gap={2}>
+          <GalleryBoardName isOpen={isBoardListOpen} onToggle={onToggleBoardList} />
+          <GallerySettingsPopover />
+        </Flex>
         <Box>
-         <BoardsListWithPrivate isOpen={isBoardListOpen} />
-         </Box>
-      ) : (
-        <Box>
-          <Flex alignItems="center" justifyContent="space-between" gap={2}>
-            <GalleryBoardName isOpen={isBoardListOpen} onToggle={onToggleBoardList} />
-            <GallerySettingsPopover />
-          </Flex>
-          <Box>
-            <BoardsList isOpen={isBoardListOpen} />
-          </Box>
+          <BoardsList isOpen={isBoardListOpen} />
         </Box>
-      )}
+      </Box>
       <Flex alignItems="center" justifyContent="space-between" gap={2}>
         <Tabs index={galleryView === 'images' ? 0 : 1} variant="unstyled" size="sm" w="full">
           <TabList>
