+++ invokeai/frontend/web/src/services/api/schema.ts
@@ -7293,145 +7293,145 @@ export type components = {
       project_id: string | null;
     };
     InvocationOutputMap: {
-      save_image: components["schemas"]["ImageOutput"];
-      integer_math: components["schemas"]["IntegerOutput"];
-      segment_anything_processor: components["schemas"]["ImageOutput"];
-      sdxl_refiner_compel_prompt: components["schemas"]["ConditioningOutput"];
-      zoe_depth_image_processor: components["schemas"]["ImageOutput"];
-      collect: components["schemas"]["CollectInvocationOutput"];
-      range: components["schemas"]["IntegerCollectionOutput"];
-      unsharp_mask: components["schemas"]["ImageOutput"];
-      string_replace: components["schemas"]["StringOutput"];
-      face_identifier: components["schemas"]["ImageOutput"];
-      heuristic_resize: components["schemas"]["ImageOutput"];
       range_of_size: components["schemas"]["IntegerCollectionOutput"];
-      latents_collection: components["schemas"]["LatentsCollectionOutput"];
-      color_map_image_processor: components["schemas"]["ImageOutput"];
-      img_ilerp: components["schemas"]["ImageOutput"];
-      infill_patchmatch: components["schemas"]["ImageOutput"];
-      face_off: components["schemas"]["FaceOffOutput"];
-      string_collection: components["schemas"]["StringCollectionOutput"];
-      sdxl_model_loader: components["schemas"]["SDXLModelLoaderOutput"];
-      sdxl_lora_loader: components["schemas"]["SDXLLoRALoaderOutput"];
-      string_join: components["schemas"]["StringOutput"];
-      lblend: components["schemas"]["LatentsOutput"];
-      conditioning_collection: components["schemas"]["ConditioningCollectionOutput"];
-      string_split_neg: components["schemas"]["StringPosNegOutput"];
-      img_watermark: components["schemas"]["ImageOutput"];
-      infill_lama: components["schemas"]["ImageOutput"];
-      div: components["schemas"]["IntegerOutput"];
-      show_image: components["schemas"]["ImageOutput"];
-      tile_to_properties: components["schemas"]["TileToPropertiesOutput"];
-      sub: components["schemas"]["IntegerOutput"];
-      normalbae_image_processor: components["schemas"]["ImageOutput"];
-      invert_tensor_mask: components["schemas"]["MaskOutput"];
-      create_gradient_mask: components["schemas"]["GradientMaskOutput"];
-      string_split: components["schemas"]["String2Output"];
-      step_param_easing: components["schemas"]["FloatCollectionOutput"];
-      metadata: components["schemas"]["MetadataOutput"];
-      img_pad_crop: components["schemas"]["ImageOutput"];
-      integer: components["schemas"]["IntegerOutput"];
-      img_mul: components["schemas"]["ImageOutput"];
-      calculate_image_tiles: components["schemas"]["CalculateImageTilesOutput"];
-      color: components["schemas"]["ColorOutput"];
-      infill_rgba: components["schemas"]["ImageOutput"];
-      t2i_adapter: components["schemas"]["T2IAdapterOutput"];
-      denoise_latents: components["schemas"]["LatentsOutput"];
-      img_lerp: components["schemas"]["ImageOutput"];
-      img_channel_offset: components["schemas"]["ImageOutput"];
-      img_crop: components["schemas"]["ImageOutput"];
-      alpha_mask_to_tensor: components["schemas"]["MaskOutput"];
-      color_correct: components["schemas"]["ImageOutput"];
-      calculate_image_tiles_min_overlap: components["schemas"]["CalculateImageTilesOutput"];
-      img_hue_adjust: components["schemas"]["ImageOutput"];
-      lresize: components["schemas"]["LatentsOutput"];
-      img_blur: components["schemas"]["ImageOutput"];
-      compel: components["schemas"]["ConditioningOutput"];
-      sdxl_lora_collection_loader: components["schemas"]["SDXLLoRALoaderOutput"];
-      float_to_int: components["schemas"]["IntegerOutput"];
-      boolean: components["schemas"]["BooleanOutput"];
-      string_join_three: components["schemas"]["StringOutput"];
-      add: components["schemas"]["IntegerOutput"];
-      merge_tiles_to_image: components["schemas"]["ImageOutput"];
-      core_metadata: components["schemas"]["MetadataOutput"];
-      lscale: components["schemas"]["LatentsOutput"];
-      mlsd_image_processor: components["schemas"]["ImageOutput"];
-      image_collection: components["schemas"]["ImageCollectionOutput"];
-      crop_latents: components["schemas"]["LatentsOutput"];
-      image_mask_to_tensor: components["schemas"]["MaskOutput"];
-      lora_collection_loader: components["schemas"]["LoRALoaderOutput"];
-      ip_adapter: components["schemas"]["IPAdapterOutput"];
-      pidi_image_processor: components["schemas"]["ImageOutput"];
-      rand_int: components["schemas"]["IntegerOutput"];
-      img_conv: components["schemas"]["ImageOutput"];
-      scheduler: components["schemas"]["SchedulerOutput"];
-      img_paste: components["schemas"]["ImageOutput"];
-      noise: components["schemas"]["NoiseOutput"];
-      img_scale: components["schemas"]["ImageOutput"];
-      i2l: components["schemas"]["LatentsOutput"];
-      main_model_loader: components["schemas"]["ModelLoaderOutput"];
-      blank_image: components["schemas"]["ImageOutput"];
-      mask_edge: components["schemas"]["ImageOutput"];
-      seamless: components["schemas"]["SeamlessModeOutput"];
-      esrgan: components["schemas"]["ImageOutput"];
-      canvas_paste_back: components["schemas"]["ImageOutput"];
-      mul: components["schemas"]["IntegerOutput"];
-      dynamic_prompt: components["schemas"]["StringCollectionOutput"];
-      controlnet: components["schemas"]["ControlOutput"];
-      l2i: components["schemas"]["ImageOutput"];
-      ideal_size: components["schemas"]["IdealSizeOutput"];
-      latents: components["schemas"]["LatentsOutput"];
-      midas_depth_image_processor: components["schemas"]["ImageOutput"];
-      tomask: components["schemas"]["ImageOutput"];
-      float_math: components["schemas"]["FloatOutput"];
-      round_float: components["schemas"]["FloatOutput"];
-      cv_inpaint: components["schemas"]["ImageOutput"];
-      create_denoise_mask: components["schemas"]["DenoiseMaskOutput"];
-      model_identifier: components["schemas"]["ModelIdentifierOutput"];
-      pair_tile_image: components["schemas"]["PairTileImageOutput"];
-      lineart_image_processor: components["schemas"]["ImageOutput"];
-      img_nsfw: components["schemas"]["ImageOutput"];
-      infill_cv2: components["schemas"]["ImageOutput"];
-      clip_skip: components["schemas"]["CLIPSkipInvocationOutput"];
-      dw_openpose_image_processor: components["schemas"]["ImageOutput"];
-      img_resize: components["schemas"]["ImageOutput"];
-      iterate: components["schemas"]["IterateInvocationOutput"];
-      rectangle_mask: components["schemas"]["MaskOutput"];
-      canny_image_processor: components["schemas"]["ImageOutput"];
-      calculate_image_tiles_even_split: components["schemas"]["CalculateImageTilesOutput"];
-      mask_from_id: components["schemas"]["ImageOutput"];
-      metadata_item: components["schemas"]["MetadataItemOutput"];
-      infill_tile: components["schemas"]["ImageOutput"];
-      tiled_multi_diffusion_denoise_latents: components["schemas"]["LatentsOutput"];
       img_channel_multiply: components["schemas"]["ImageOutput"];
-      boolean_collection: components["schemas"]["BooleanCollectionOutput"];
-      lora_loader: components["schemas"]["LoRALoaderOutput"];
-      float_collection: components["schemas"]["FloatCollectionOutput"];
-      string: components["schemas"]["StringOutput"];
-      freeu: components["schemas"]["UNetOutput"];
-      lineart_anime_image_processor: components["schemas"]["ImageOutput"];
-      depth_anything_image_processor: components["schemas"]["ImageOutput"];
-      image: components["schemas"]["ImageOutput"];
-      face_mask_detection: components["schemas"]["FaceMaskOutput"];
-      rand_float: components["schemas"]["FloatOutput"];
-      float: components["schemas"]["FloatOutput"];
-      random_range: components["schemas"]["IntegerCollectionOutput"];
-      integer_collection: components["schemas"]["IntegerCollectionOutput"];
-      sdxl_refiner_model_loader: components["schemas"]["SDXLRefinerModelLoaderOutput"];
-      mask_combine: components["schemas"]["ImageOutput"];
-      tile_image_processor: components["schemas"]["ImageOutput"];
-      img_chan: components["schemas"]["ImageOutput"];
-      vae_loader: components["schemas"]["VAEOutput"];
-      prompt_from_file: components["schemas"]["StringCollectionOutput"];
-      float_range: components["schemas"]["FloatCollectionOutput"];
+      img_blur: components["schemas"]["ImageOutput"];
+      cv_inpaint: components["schemas"]["ImageOutput"];
+      integer_math: components["schemas"]["IntegerOutput"];
       merge_metadata: components["schemas"]["MetadataOutput"];
-      sdxl_compel_prompt: components["schemas"]["ConditioningOutput"];
-      hed_image_processor: components["schemas"]["ImageOutput"];
-      lora_selector: components["schemas"]["LoRASelectorOutput"];
-      conditioning: components["schemas"]["ConditioningOutput"];
+      mask_combine: components["schemas"]["ImageOutput"];
+      sub: components["schemas"]["IntegerOutput"];
+      crop_latents: components["schemas"]["LatentsOutput"];
+      ideal_size: components["schemas"]["IdealSizeOutput"];
       leres_image_processor: components["schemas"]["ImageOutput"];
+      mask_from_id: components["schemas"]["ImageOutput"];
+      infill_cv2: components["schemas"]["ImageOutput"];
+      boolean: components["schemas"]["BooleanOutput"];
+      conditioning_collection: components["schemas"]["ConditioningCollectionOutput"];
+      infill_patchmatch: components["schemas"]["ImageOutput"];
+      float_math: components["schemas"]["FloatOutput"];
+      image_collection: components["schemas"]["ImageCollectionOutput"];
+      img_conv: components["schemas"]["ImageOutput"];
+      sdxl_refiner_compel_prompt: components["schemas"]["ConditioningOutput"];
+      metadata: components["schemas"]["MetadataOutput"];
+      save_image: components["schemas"]["ImageOutput"];
+      boolean_collection: components["schemas"]["BooleanCollectionOutput"];
+      img_paste: components["schemas"]["ImageOutput"];
+      conditioning: components["schemas"]["ConditioningOutput"];
+      color_map_image_processor: components["schemas"]["ImageOutput"];
+      img_scale: components["schemas"]["ImageOutput"];
+      string_join: components["schemas"]["StringOutput"];
+      noise: components["schemas"]["NoiseOutput"];
+      core_metadata: components["schemas"]["MetadataOutput"];
+      lresize: components["schemas"]["LatentsOutput"];
+      face_mask_detection: components["schemas"]["FaceMaskOutput"];
+      calculate_image_tiles_even_split: components["schemas"]["CalculateImageTilesOutput"];
+      mlsd_image_processor: components["schemas"]["ImageOutput"];
+      sdxl_lora_collection_loader: components["schemas"]["SDXLLoRALoaderOutput"];
+      img_nsfw: components["schemas"]["ImageOutput"];
+      main_model_loader: components["schemas"]["ModelLoaderOutput"];
+      rand_int: components["schemas"]["IntegerOutput"];
+      mul: components["schemas"]["IntegerOutput"];
+      integer: components["schemas"]["IntegerOutput"];
+      img_pad_crop: components["schemas"]["ImageOutput"];
+      collect: components["schemas"]["CollectInvocationOutput"];
+      pidi_image_processor: components["schemas"]["ImageOutput"];
+      lora_loader: components["schemas"]["LoRALoaderOutput"];
+      infill_tile: components["schemas"]["ImageOutput"];
+      calculate_image_tiles: components["schemas"]["CalculateImageTilesOutput"];
+      scheduler: components["schemas"]["SchedulerOutput"];
+      midas_depth_image_processor: components["schemas"]["ImageOutput"];
+      latents: components["schemas"]["LatentsOutput"];
+      tile_to_properties: components["schemas"]["TileToPropertiesOutput"];
+      round_float: components["schemas"]["FloatOutput"];
+      add: components["schemas"]["IntegerOutput"];
+      dynamic_prompt: components["schemas"]["StringCollectionOutput"];
+      l2i: components["schemas"]["ImageOutput"];
+      integer_collection: components["schemas"]["IntegerCollectionOutput"];
+      model_identifier: components["schemas"]["ModelIdentifierOutput"];
+      normalbae_image_processor: components["schemas"]["ImageOutput"];
+      img_chan: components["schemas"]["ImageOutput"];
+      image: components["schemas"]["ImageOutput"];
+      face_off: components["schemas"]["FaceOffOutput"];
+      img_hue_adjust: components["schemas"]["ImageOutput"];
+      color: components["schemas"]["ColorOutput"];
+      color_correct: components["schemas"]["ImageOutput"];
+      string_split_neg: components["schemas"]["StringPosNegOutput"];
+      clip_skip: components["schemas"]["CLIPSkipInvocationOutput"];
+      face_identifier: components["schemas"]["ImageOutput"];
+      string_join_three: components["schemas"]["StringOutput"];
+      img_channel_offset: components["schemas"]["ImageOutput"];
+      infill_rgba: components["schemas"]["ImageOutput"];
+      show_image: components["schemas"]["ImageOutput"];
+      infill_lama: components["schemas"]["ImageOutput"];
+      img_mul: components["schemas"]["ImageOutput"];
+      segment_anything_processor: components["schemas"]["ImageOutput"];
+      sdxl_refiner_model_loader: components["schemas"]["SDXLRefinerModelLoaderOutput"];
       mediapipe_face_processor: components["schemas"]["ImageOutput"];
+      image_mask_to_tensor: components["schemas"]["MaskOutput"];
+      string_split: components["schemas"]["String2Output"];
+      i2l: components["schemas"]["LatentsOutput"];
+      zoe_depth_image_processor: components["schemas"]["ImageOutput"];
+      float_to_int: components["schemas"]["IntegerOutput"];
+      ip_adapter: components["schemas"]["IPAdapterOutput"];
+      string: components["schemas"]["StringOutput"];
+      rectangle_mask: components["schemas"]["MaskOutput"];
+      lineart_anime_image_processor: components["schemas"]["ImageOutput"];
+      alpha_mask_to_tensor: components["schemas"]["MaskOutput"];
+      merge_tiles_to_image: components["schemas"]["ImageOutput"];
+      div: components["schemas"]["IntegerOutput"];
+      float_collection: components["schemas"]["FloatCollectionOutput"];
+      tiled_multi_diffusion_denoise_latents: components["schemas"]["LatentsOutput"];
+      seamless: components["schemas"]["SeamlessModeOutput"];
+      float: components["schemas"]["FloatOutput"];
+      lineart_image_processor: components["schemas"]["ImageOutput"];
+      controlnet: components["schemas"]["ControlOutput"];
+      step_param_easing: components["schemas"]["FloatCollectionOutput"];
+      prompt_from_file: components["schemas"]["StringCollectionOutput"];
+      rand_float: components["schemas"]["FloatOutput"];
+      float_range: components["schemas"]["FloatCollectionOutput"];
+      sdxl_lora_loader: components["schemas"]["SDXLLoRALoaderOutput"];
+      latents_collection: components["schemas"]["LatentsCollectionOutput"];
+      sdxl_compel_prompt: components["schemas"]["ConditioningOutput"];
+      create_denoise_mask: components["schemas"]["DenoiseMaskOutput"];
+      create_gradient_mask: components["schemas"]["GradientMaskOutput"];
+      metadata_item: components["schemas"]["MetadataItemOutput"];
+      img_ilerp: components["schemas"]["ImageOutput"];
       content_shuffle_image_processor: components["schemas"]["ImageOutput"];
+      img_watermark: components["schemas"]["ImageOutput"];
+      invert_tensor_mask: components["schemas"]["MaskOutput"];
+      img_crop: components["schemas"]["ImageOutput"];
+      esrgan: components["schemas"]["ImageOutput"];
+      iterate: components["schemas"]["IterateInvocationOutput"];
+      unsharp_mask: components["schemas"]["ImageOutput"];
+      mask_edge: components["schemas"]["ImageOutput"];
+      canvas_paste_back: components["schemas"]["ImageOutput"];
+      dw_openpose_image_processor: components["schemas"]["ImageOutput"];
+      tomask: components["schemas"]["ImageOutput"];
+      random_range: components["schemas"]["IntegerCollectionOutput"];
+      lora_selector: components["schemas"]["LoRASelectorOutput"];
+      compel: components["schemas"]["ConditioningOutput"];
+      img_resize: components["schemas"]["ImageOutput"];
+      sdxl_model_loader: components["schemas"]["SDXLModelLoaderOutput"];
+      lora_collection_loader: components["schemas"]["LoRALoaderOutput"];
+      string_replace: components["schemas"]["StringOutput"];
+      string_collection: components["schemas"]["StringCollectionOutput"];
+      blank_image: components["schemas"]["ImageOutput"];
+      freeu: components["schemas"]["UNetOutput"];
+      img_lerp: components["schemas"]["ImageOutput"];
+      range: components["schemas"]["IntegerCollectionOutput"];
+      depth_anything_image_processor: components["schemas"]["ImageOutput"];
+      t2i_adapter: components["schemas"]["T2IAdapterOutput"];
+      pair_tile_image: components["schemas"]["PairTileImageOutput"];
+      lscale: components["schemas"]["LatentsOutput"];
+      tile_image_processor: components["schemas"]["ImageOutput"];
+      heuristic_resize: components["schemas"]["ImageOutput"];
+      calculate_image_tiles_min_overlap: components["schemas"]["CalculateImageTilesOutput"];
+      denoise_latents: components["schemas"]["LatentsOutput"];
+      vae_loader: components["schemas"]["VAEOutput"];
+      lblend: components["schemas"]["LatentsOutput"];
+      hed_image_processor: components["schemas"]["ImageOutput"];
+      canny_image_processor: components["schemas"]["ImageOutput"];
     };
     /**
      * InvocationStartedEvent
@@ -15014,6 +15014,8 @@ export type operations = {
       query: {
         /** @description The name of the board to create */
         board_name: string;
+        /** @description Whether the board is private */
+        private_board?: boolean;
       };
     };
     responses: {
