+++ invokeai/frontend/web/src/services/api/types.ts
@@ -11,9 +11,7 @@ export type ListBoardsArgs = NonNullable<paths['/api/v1/boards/']['get']['parame
 export type DeleteBoardResult =
   paths['/api/v1/boards/{board_id}']['delete']['responses']['200']['content']['application/json'];
 
-export type CreateBoardArg = paths['/api/v1/boards/']['post']['parameters']['query'] & {
-  changes: paths['/api/v1/boards/']['post']['parameters']['query'];
-};
+export type CreateBoardArg = paths['/api/v1/boards/']['post']['parameters']['query'];
 
 export type UpdateBoardArg = paths['/api/v1/boards/{board_id}']['patch']['parameters']['path'] & {
   changes: paths['/api/v1/boards/{board_id}']['patch']['requestBody']['content']['application/json'];
